+++ openai-client/src/commonTest/kotlin/com/aallam/openai/client/TestOpenAI.kt
@@ -53,7 +53,6 @@ class TestOpenAI {
                 temperature = 1.0,
                 topP = 1.0,
                 n = 1,
-                stream = false,
                 logprobs = null,
                 stop = listOf("\n")
             )
@@ -72,7 +71,6 @@ class TestOpenAI {
                 temperature = 1.0,
                 topP = 1.0,
                 n = 1,
-                stream = true,
                 logprobs = null,
                 stop = listOf("\n"),
             )